### 2018-05-06 #725 @whedon
[PRE REVIEW]: mbir: Magnitude-Based Inferences
**Submitting author:** @kdpeterson51 (<a href="http://orcid.org/0000-0003-0128-5282">Kyle Peterson</a>)
**Repository:** <a href="https://github.com/kdpeterson51/mbir" target ="_blank">https://github.com/kdpeterson51/mbir</a>
**Version:** 1.3
