### 2018-05-06 #725 @arfon
@kdpeterson51 - please take a look at the guidelines for the paper structure http://joss.theoj.org/about#paper_structure and update yours accordingly.

In addition, this package doesn't look like it has a README, license etc. You might want to take a look at the [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines) to see what your package will be evaluated on.
