+++ README.md
@@ -1,1 +1,2 @@
+[![version](http://www.r-pkg.org/badges/version/mbir)](https://cran.r-project.org/package=mbir)
 [![CRAN](https://cranlogs.r-pkg.org/badges/mbir)](https://cran.r-project.org/web/packages/mbir/index.html)
