+++ README.md
@@ -2,3 +2,5 @@
 
 [![version](http://www.r-pkg.org/badges/version/mbir)](https://cran.r-project.org/package=mbir)
 [![CRAN](https://cranlogs.r-pkg.org/badges/mbir)](https://cran.r-project.org/web/packages/mbir/index.html)
+
+Allows practitioners and researchers a wholesale approach for deriving magnitude-based inferences from raw data. A major goal of 'mbir' is to programmatically detect appropriate statistical tests to run in lieu of relying on practitioners to determine correct stepwise procedures independently.
