+++ README.md
@@ -5,6 +5,11 @@
 Kyle D Peterson and Aaron R Caldwell
 
 # Overview
-Allows practitioners and researchers a wholesale approach for deriving magnitude-based inferences from raw data. A major goal of 'mbir' is to programmatically detect appropriate statistical tests to run in lieu of relying on practitioners to determine correct stepwise procedures independently.
+Allows practitioners and researchers a wholesale approach for deriving magnitude-based inferences from raw data. A major goal of `mbir` is to programmatically detect appropriate statistical tests to run in lieu of relying on practitioners to determine correct stepwise procedures independently.
 
 # Installation
+Assuming users have R downloaded and installed, `mbir` can be installed right from [CRAN](https://cran.r-project.org/web/packages/mbir/index.html) by typing: 
+
+```
+install.packages("mbir")
+```
