+++ README.md
@@ -1,7 +1,7 @@
 [![version](http://www.r-pkg.org/badges/version/mbir)](https://cran.r-project.org/package=mbir)
 [![CRAN](https://cranlogs.r-pkg.org/badges/mbir)](https://cran.r-project.org/web/packages/mbir/index.html)
 
-# mbir
+# mbir: Magnitude-Based Inferences
 Kyle D Peterson and Aaron R Caldwell
 
 # Overview
