+++ README.md
@@ -15,3 +15,5 @@ install.packages("mbir")
 ```
 This package depends on the packages `graphics`, `stats`, and `utils`, which are imported automatically upon installation.
 
+# Feedback
+Feedback from users is welcome, and would be sincerely appreciated, to help improve functionality of `mbir` where warranted. Please reach out to petersonkdon@gmail.com for support, reporting issues, or contributions. Thank you very much.
