### 2018-05-06 0eac8cb
Moving files to correct location and adding R functions.
