+++ mbir/R/datambinpind.b.R
@@ -1,0 +1,294 @@
+
+# This file is a generated template, your changes will not be overwritten
+
+dataMBInpindClass <- if (requireNamespace('jmvcore')) R6::R6Class(
+    "dataMBInpindClass",
+    inherit = dataMBInpindBase,
+    private = list(
+      .init = function() {
+        
+        
+        table <- self$results$ttest
+        EffSize <- self$results$effect
+        MBI <- self$results$mbi
+        desc <- self$results$desc
+        
+        
+        EffSize$getColumn('rLL')$setSuperTitle(jmvcore::format('{}% Confidence Interval', self$options$confint))
+        EffSize$getColumn('rUL')$setSuperTitle(jmvcore::format('{}% Confidence Interval', self$options$confint))
+        
+        groupName <- self$options$group
+        groups <- NULL
+        if ( ! is.null(groupName))
+          groups <- base::levels(self$data[[groupName]])
+        if (length(groups) != 2)
+          groups <- c('Group 1', 'Group 2')
+        
+        desc <- self$results$desc
+        for (key in desc$rowKeys) {
+          desc$setRow(rowKey=key, values=list(
+            `name[1]`=groups[1],
+            `name[2]`=groups[2]))
+        }
+        
+ 
+        
+      },
+        .run = function() {
+          
+          if (is.null(self$options$group) || length(self$options$deps) == 0)
+            return()
+          
+          table <- self$results$ttest
+          EffSize <- self$results$effect
+          MBI <- self$results$mbi
+          desc <- self$results$desc
+          
+          conf.int <- (self$options$confint)/100
+          SWC <- self$options$SWC
+          
+          groupName <- self$options$group
+          group <- self$data[[groupName]]
+          group <- as.factor(group)
+          group <- droplevels(group)
+          
+          groupLevels <- base::levels(group)
+          if (length(groupLevels) != 2)
+            jmvcore::reject("Grouping variable must have exactly 2 levels", code="grouping_var_must_have_2_levels")
+          
+          
+          for (depName in self$options$deps) {
+            
+            
+            dep <- self$data[[depName]]
+            dep <- jmvcore::toNumeric(dep)
+            dataTTest <- data.frame(dep=dep, group=group)
+            dataTTest <- na.omit(dataTTest)
+            
+            n <- tapply(dataTTest$dep, dataTTest$group, length)
+            n1 <- n[1]
+            n2 <- n[2]
+            n3 <- n1+n2
+            
+            v <- tapply(dataTTest$dep, dataTTest$group, function(x) jmvcore::tryNaN(var(x)))
+            v1 <- v[1]
+            v2 <- v[2]
+            
+            m <- tapply(dataTTest$dep, dataTTest$group, function(x) jmvcore::tryNaN(mean(x)))
+            m1 <- m[1]
+            m2 <- m[2]
+            
+            med <- tapply(dataTTest$dep, dataTTest$group, function(x) jmvcore::tryNaN(median(x)))
+            med1 <- med[1]
+            med2 <- med[2]
+            
+            se <- sqrt(v/n)
+            se1 <- se[1]
+            se2 <- se[2]
+            
+            sd <- sqrt(v)
+            sd1 <- sd[1]
+            sd2 <- sd[2]
+            
+            
+            rank <- stats::wilcox.test(dep ~ group, dataTTest,
+              paired = FALSE,
+              conf.int = T,
+              conf.level = conf.int,
+              na.action = na.omit,
+              correct = F,
+              exact = F
+            )
+            
+            suppressWarnings(warning(rank))
+            diff <- rank$estimate
+            
+            r <- ifelse(diff < 0,
+                        stats::qnorm(rank$p.value / 2) / sqrt(n3),
[…]
