+++ mbir/R/datambinpind.h.R
@@ -1,0 +1,310 @@
+
+# This file is automatically generated, you probably don't want to edit this
+
+dataMBInpindOptions <- if (requireNamespace('jmvcore')) R6::R6Class(
+    "dataMBInpindOptions",
+    inherit = jmvcore::Options,
+    public = list(
+        initialize = function(
+            deps = NULL,
+            group = NULL,
+            SWC = 0.5,
+            confint = 90,
+            desc = FALSE, ...) {
+
+            super$initialize(
+                package='mbir',
+                name='dataMBInpind',
+                requiresData=TRUE,
+                ...)
+
+            private$..deps <- jmvcore::OptionVariables$new(
+                "deps",
+                deps,
+                suggested=list(
+                    "continuous"),
+                permitted=list(
+                    "continuous",
+                    "nominal",
+                    "ordinal"),
+                rejectInf=FALSE)
+            private$..group <- jmvcore::OptionVariable$new(
+                "group",
+                group,
+                suggested=list(
+                    "nominal",
+                    "ordinal"))
+            private$..SWC <- jmvcore::OptionNumber$new(
+                "SWC",
+                SWC,
+                min=0,
+                default=0.5)
+            private$..confint <- jmvcore::OptionNumber$new(
+                "confint",
+                confint,
+                min=50,
+                max=99.9,
+                default=90)
+            private$..desc <- jmvcore::OptionBool$new(
+                "desc",
+                desc,
+                default=FALSE)
+
+            self$.addOption(private$..deps)
+            self$.addOption(private$..group)
+            self$.addOption(private$..SWC)
+            self$.addOption(private$..confint)
+            self$.addOption(private$..desc)
+        }),
+    active = list(
+        deps = function() private$..deps$value,
+        group = function() private$..group$value,
+        SWC = function() private$..SWC$value,
+        confint = function() private$..confint$value,
+        desc = function() private$..desc$value),
+    private = list(
+        ..deps = NA,
+        ..group = NA,
+        ..SWC = NA,
+        ..confint = NA,
+        ..desc = NA)
+)
+
+dataMBInpindResults <- if (requireNamespace('jmvcore')) R6::R6Class(
+    inherit = jmvcore::Group,
+    active = list(
+        text = function() private$.items[["text"]],
+        ttest = function() private$.items[["ttest"]],
+        effect = function() private$.items[["effect"]],
+        mbi = function() private$.items[["mbi"]],
+        desc = function() private$.items[["desc"]]),
+    private = list(),
+    public=list(
+        initialize=function(options) {
+            super$initialize(
+                options=options,
+                name="",
+                title="MBI for Independent Samples (non-parametric)")
+            self$add(jmvcore::Preformatted$new(
+                options=options,
+                name="text",
+                title="MBI for Mann-Whitney"))
+            self$add(jmvcore::Table$new(
+                options=options,
+                name="ttest",
+                title="Mann-Whitney U-Test",
+                rows="(deps)",
+                columns=list(
+                    list(
+                        `name`="var", 
+                        `title`="", 
+                        `type`="text", 
+                        `content`="($key)"),
+                    list(
+                        `name`="statistic", 
+                        `type`="number"),
+                    list(
+                        `name`="p", 
+                        `type`="number", 
+                        `format`="zto,pvalue"))))
+            self$add(jmvcore::Table$new(
+                options=options,
+                name="effect",
+                title="Effect Size",
+                rows="(deps)",
+                clearWith=list(
+                    "confint",
+                    "SWC"),
+                columns=list(
+                    list(
+                        `name`="var", 
+                        `title`="", 
+                        `type`="text", 
[…]

[…]
