+++ .travis.yml
@@ -1,2 +1,3 @@
+sudo: false
 language: r
 cache: packages
