### 2018-05-06 #725 @kdpeterson51
Sincere apologies, Afron.

Tags have been added  to YAML header, README and LICENSE files are now up-to-date.
