+++ README.md
@@ -28,6 +28,6 @@ Below is an example of how a user would perform a correlation with the classic m
 data('mtcars')
 mbir::corr_test(mtcars$mpg,mtcars$qsec)
 ```
-
+For a detailed exporlation of `mbir`, please visit our [vignette](http://www.mbir-project.us/mbir.pdf).
 # Feedback
 Feedback from users is welcome, and would be sincerely appreciated, to help improve functionality of `mbir` where warranted. Please reach out to petersonkdon@gmail.com for support, reporting issues, or contributions. Thank you very much.
