### 2018-05-09 #725 @leeper
Hi, @kdpeterson51. Thanks for your submission. A couple of preliminary things here before we get started:

1. Can you clarify the license for the functions mentioned in [COPYRIGHT](https://github.com/kdpeterson51/mbir/blob/master/mbir/inst/COPYRIGHT)? Do you have permission to redistribute this code from the original author? If so, under what license?

2. It looks like you are generating your R [NAMESPACE file](https://github.com/kdpeterson51/mbir/blob/master/mbir/NAMESPACE) by hand rather than through roxygen. Can you delete and reinitialize using `devtools::document()`?

3. Should [the vignette](http://www.mbir-project.us/mbir.pdf) be included in a typical `/vignettes` directory in the package itself?

4. Do you have any suggested reviewers for the package?
