+++ mbir/vignettes/mbir-vignette.Rmd
@@ -1,0 +1,486 @@
+---
+title: "Introduction to `mbir`"
+author: "Kyle D. Peterson, Aaron Caldwell"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Introduction to mbir}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+This vignette is intended to demonstrate the basic usage of `mbir` (**M**agnitude-**B**ased **I**nferences via **R**).
+
+## Inspiration
+Before we begin, users must be aware that `mbir` was originally inspired by the work of Will Hopkins^[Hopkins WG. sportsci.org.], and intended to inaugurate his work to the **R** community. Therefore, to give credit where credit is due, this vignette will formally declare which segments of `mbir` were directly derived from Hopkins.   
+
+Also, the purpose of `mbir` is to provide practitioners, and researchers, a wholesale approach for deriving magnitude-based inferences from raw data. That is, in lieu of relying on practitioners to determine correct stepwise procedures independently, `mbir` programmatically detects appropriate statistical test(s) to run per the given data. This utility may serve as an educational tool to some (or perhaps alleviate doubt), but most importantly, will deliver a consistent platform for **R** users who aspire to employ this type of analysis. Respective functionalities will be dissected in later sections. 
+
+## Overview
+**Section 1: Data Functions**   
+To start, `mbir` is currently equipped to analyze independent or paired, two-sample data. (Subsequent versions are planned to broaden the package's utility). Functions that end in "`_test`" are meant to accept raw data and perform statistical tests. The following functions are to investigate difference in means and associations, respectively:
+
+
+- `smd_test()`
+- `corr_test()`
+
+**Section 2: Input Functions**   
+On the contrary, functions listed below are structured to compute magnitude-based inferences for users who may have already performed statistical tests on their own:
+
+
+- `smd()`
+- `corr()`
+- `odds()`
+
+The additional functions below allows users to apply magnitude-based thinking when analyzing two proportions and two correlations, respectively:   
+
+- `prop()`
+- `corr_diff()`
+
+**Section 3: Sample Size Functions**   
+To appease the research-minded practitioners, functions that start with "`ss_`" assist with the critical decision of *a priori* sample size planning that correspond to the three input functions listed above:
+
+- `ss_smd()`
+- `ss_corr()`
+- `ss_odds()`
+
+For users who may prefer Accuracy in Parameter Estimation (AIPE) - the closely akin methodology to sample size planning via precision - can be employed with the below function:
+
+- `aipe_smd()`
+
+**Section 4: Miscellaneous Functions**   
+If perhaps users are not familiar with a given effect size measure, `es_convert()` provides the service of converting between equivalent effect size measures. 
+
+If users wish to calculate confidence intervals via resampling, `boot_test()` is a convenient function for this task. 
+
+Lastly, `swc_ind()` assesses longitudinal change of an individual's performance measurements in a linear fashion.
+
+
+## Section 1: Data Functions
+`mbir` Data Functions are designed to accept raw data and execute appropriate statistical test(s) programmatically. The thematic operation of `mbir` Data Functions follow the statistical progression of:    
+
+1. Assess Assumptions   
+2. Perform Null-Hypothesis Statistical Test(s)   
+3. Calculate Effect Size(s)   
+4. Compute Magnitude-Based Inferences   
+
+Each of these steps are obligatory for proper statistical estimations of effect, which was a major incentive for the creation of `mbir`.  Let's now examine the underlying mechanisms and implementation of `mbir` Data Functions.
+
+### 1.1. `smd_test()`
[…]
