### 2018-05-16 4a36a53
Delete NAMESPACE
