+++ mbir/NAMESPACE
@@ -1,0 +1,16 @@
+# Generated by roxygen2: do not edit by hand
+
+export(aipe_smd)
+export(boot_test)
+export(corr)
+export(corr_diff)
+export(corr_test)
+export(es_convert)
+export(odds)
+export(prop)
+export(smd)
+export(smd_test)
+export(ss_corr)
+export(ss_odds)
+export(ss_smd)
+export(swc_ind)
