### 2018-05-16 #725 @kdpeterson51
Hi @leeper. Thank you very much for your assistance.

1. The copyright was obtained by written permission to use those cited MS Excel spreadsheets.

2. Original NAMESPACE file has been deleted and replaced with `devtools` version.

3. Vignette folder is now up-to-date within `/mbir` folder.

4. We do not have any suggested reviewers at this time.
