### 2018-05-17 #725 @kdpeterson51
Your suggestions are appreciated, @leeper. The [COPYRIGHT](https://github.com/kdpeterson51/mbir/blob/master/mbir/inst/COPYRIGHT) file has been updated to improve the language.
