### 2018-05-18 #725 @leeper
@kellieotto Would you be interested in reviewing this software for JOSS?
