### 2018-05-30 #746 @kellieotto
I've done a first review. The version and automated tests need attention. I'm leaving detailed feedback in the repository issues.
