### 2018-05-30 kdpeterson51/mbir#3 @kellieotto
Missing version on Github
This is detailed feedback for your JOSS review openjournals/joss-reviews#746.

You should create a version of your package in the Github repository that matches the version number on CRAN. Versioning is tracked here: https://github.com/kdpeterson51/mbir/releases
