### 2018-06-02 #746 @leeper
@kdpeterson51 Please address the issues raised in the review and then let me know once your updates are ready. I'll then take a look and decide if further review is needed.
