+++ mbir/DESCRIPTION
@@ -1,7 +1,7 @@
 Package: mbir
 Type: Package
 Title: Magnitude-Based Inferences
-Version: 1.3
+Version: 1.3.1
 Date: 2018-04-28
 Authors@R: c(person("Kyle", "Peterson", 
                      email = "petersonkdon@gmail.com",
@@ -17,5 +17,6 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
 Suggests: knitr,
-    rmarkdown
+    rmarkdown,
+    testthat
 VignetteBuilder: knitr
