### 2018-06-03 kdpeterson51/mbir#3 @kdpeterson51
Thanks for your help, @kellieotto.

The most current version on CRAN (1.3.1) has been uploaded to Github.
