### 2018-06-03 #746 @kdpeterson51
Thank you for your patience, @leeper. One of the two issues raised by @kellieotto has been [addressed](https://github.com/kdpeterson51/mbir/issues/3).

To address the second, I am currently in progress of pairing up Travis CI and will update you immediately  upon successful integration.
