### 2018-06-05 ed0186d
Change directory to mbir
