+++ .travis.yml
@@ -1,3 +1,6 @@
 sudo: false
 language: r
 cache: packages
+
+before_install:
+  - cd mbir
