### 2018-06-13 0ec1aa1
Delete 0000.yaml
