### 2018-06-13 88c0c2a
Delete datambicorr.a.yaml
