### 2018-06-13 c87a097
Delete datambicorr.r.yaml
