### 2018-06-13 c38091b
Delete datambinpind.a.yaml
