### 2018-06-13 5449f3a
Delete datambinpind.r.yaml
