+++ mbir/jamovi/datambinpind.u.yaml
@@ -1,47 +1,0 @@
-title: MBI for Independent Samples (non-parametric)
-name: dataMBInpind
-jus: '2.0'
-stage: 0
-compilerMode: tame
-children:
-  - type: VariableSupplier
-    persistentItems: false
-    stretchFactor: 1
-    children:
-      - type: TargetLayoutBox
-        label: Dependent Variables
-        children:
-          - type: VariablesListBox
-            name: deps
-            isTarget: true
-      - type: TargetLayoutBox
-        label: Grouping Variable
-        children:
-          - type: VariablesListBox
-            name: group
-            maxItemCount: 1
-            isTarget: true
-            
-  - type: LayoutBox
-    margin: large
-    stretchFactor: 1
-    children:
-      - type: LayoutBox
-        cell:
-          column: 0
-          row: 0
-        stretchFactor: 1
-        children:
-          - type: Label
-            label: Options
-            children:  
-              - type: TextBox
-                name: SWC
-                format: number
-                inputPattern: '[0-9]+'
-              - type: TextBox
-                name: confint
-                format: number
-                inputPattern: '[0-9]+'
-              - type: CheckBox
-                name: desc
