### 2018-06-13 2cf8742
Delete datambipaired.a.yaml
