### 2018-06-13 f5c227c
Delete datambipaired.r.yaml
