### 2018-06-13 d4add88
Delete datambitwo.a.yaml
