### 2018-06-13 9d743b9
Delete datambicorr.b.R
