+++ mbir/R/datambicorr.b.R
@@ -1,292 +1,0 @@
-
-# This file is a generated template, your changes will not be overwritten
-
-dataMBIcorrClass <- if (requireNamespace('jmvcore')) R6::R6Class(
-    "dataMBIcorrClass",
-    inherit = dataMBIcorrBase,
-    private = list(
-      .init = function() {
-        
-        
-        MBI <- self$results$mbi
-        desc <- self$results$desc
-        
-        confint <- self$options$confint / 100
-        
-        
-        for (pair in self$options$pairs) {
-          MBI$setRow(rowKey=pair, list(i1=pair[[1]], i2=pair[[2]]))
-          desc$setRow(rowKey=pair, list(`name[1]`=pair[[1]], `name[2]`=pair[[2]]))
-        }
-        
-        
-
-        
-        
-        
-      },
-      
-        .run = function() {
-
-          MBI <- self$results$mbi
-          desc <- self$results$desc
-          
-          
-          conf.int <- (self$options$confint)/100
-          SWC <- self$options$SWC
-          threshold <- (0.5*log((1+SWC)/(1-SWC)))         
-          
-          for (pair in self$options$pairs) {
-            
-            if (is.null(pair[[1]]))
-              next()
-            if (is.null(pair[[2]]))
-              next()
-            
-            i1 <- jmvcore::toNumeric(self$data[[ pair[[1]] ]])
-            i2 <- jmvcore::toNumeric(self$data[[ pair[[2]] ]])
-            data <- data.frame(i1=i1, i2=i2)
-            data <- na.omit(data)
-            n <- nrow(data)
-            i1 <- data$i1
-            i2 <- data$i2
-            m1 <- base::mean(i1)
-            m2 <- base::mean(i2)
-            med1 <- stats::median(i1)
-            med2 <- stats::median(i2)
-            sd1  <- stats::sd(i1)
-            sd2  <- stats::sd(i2)
-            se1  <- sd1/sqrt(n)
-            se2  <- sd2/sqrt(n)
-            
-            if (self$options$spear) {
-              
-            result <- stats::cor.test(i1, i2, method="spearman", exact=F, conf.level=conf.int)
-            
-            type_Spear <- paste("Spearman")
-            type_Spear2 <- paste("rho =")
-            p_Spear <- unname(result$p.value)
-            correl_Spear <- unname(result$estimate)
-            z_Spear <- (0.5 * 
-                         log((1 + correl_Spear)/(1 - correl_Spear)))
-            
-            zSE_Spear <- (1/sqrt(n-3))
-            
-            Positive_Spear <- round(100 * (1 - stats::pnorm(threshold, mean = z_Spear, sd = (1/sqrt(n - 3)))), digits = 1)
-            Negative_Spear <- round(100 * (stats::pnorm(-threshold, mean = z_Spear, sd = (1/sqrt(n - 3)))), digits = 1)
-            Trivial_Spear <- round(100 - Positive_Spear - Negative_Spear, digits = 1)
-            
-            
-            lower <- ifelse(Negative_Spear < 0.5, "Most Unlikely", ifelse(Negative_Spear < 
-                                                                           5, "Very Unlikely", ifelse(Negative_Spear < 25, "Unlikely", 
-                                                                                                      ifelse(Negative_Spear < 75, "Possibly", ifelse(Negative_Spear < 95, 
-                                                                                                                                                    "Likely", ifelse(Negative_Spear < 99, "Most Likely", ifelse(Negative_Spear >= 
-                                                                                                                                                                                                                 99, "Almost Certainly")))))))
-            trivial2 <- ifelse(Trivial_Spear < 0.5, "Most Unlikely", ifelse(Trivial_Spear < 
-                                                                             5, "Very Unlikely", ifelse(Trivial_Spear < 25, "Unlikely", 
[…]
