+++ mbir/R/datambicorr.h.R
@@ -1,396 +1,0 @@
-
-# This file is automatically generated, you probably don't want to edit this
-
-dataMBIcorrOptions <- if (requireNamespace('jmvcore')) R6::R6Class(
-    "dataMBIcorrOptions",
-    inherit = jmvcore::Options,
-    public = list(
-        initialize = function(
-            pairs = NULL,
-            pearson = TRUE,
-            spear = FALSE,
-            kendall = FALSE,
-            SWC = 0.1,
-            confint = 90,
-            desc = FALSE, ...) {
-
-            super$initialize(
-                package='mbir',
-                name='dataMBIcorr',
-                requiresData=TRUE,
-                ...)
-
-            private$..pairs <- jmvcore::OptionPairs$new(
-                "pairs",
-                pairs,
-                suggested=list(
-                    "continuous"),
-                permitted=list(
-                    "continuous",
-                    "nominal",
-                    "ordinal"))
-            private$..pearson <- jmvcore::OptionBool$new(
-                "pearson",
-                pearson,
-                default=TRUE)
-            private$..spear <- jmvcore::OptionBool$new(
-                "spear",
-                spear,
-                default=FALSE)
-            private$..kendall <- jmvcore::OptionBool$new(
-                "kendall",
-                kendall,
-                default=FALSE)
-            private$..SWC <- jmvcore::OptionNumber$new(
-                "SWC",
-                SWC,
-                min=0.01,
-                max=0.9,
-                default=0.1)
-            private$..confint <- jmvcore::OptionNumber$new(
-                "confint",
-                confint,
-                min=50,
-                max=99.9,
-                default=90)
-            private$..desc <- jmvcore::OptionBool$new(
-                "desc",
-                desc,
-                default=FALSE)
-
-            self$.addOption(private$..pairs)
-            self$.addOption(private$..pearson)
-            self$.addOption(private$..spear)
-            self$.addOption(private$..kendall)
-            self$.addOption(private$..SWC)
-            self$.addOption(private$..confint)
-            self$.addOption(private$..desc)
-        }),
-    active = list(
-        pairs = function() private$..pairs$value,
-        pearson = function() private$..pearson$value,
-        spear = function() private$..spear$value,
-        kendall = function() private$..kendall$value,
-        SWC = function() private$..SWC$value,
-        confint = function() private$..confint$value,
-        desc = function() private$..desc$value),
-    private = list(
-        ..pairs = NA,
-        ..pearson = NA,
-        ..spear = NA,
-        ..kendall = NA,
-        ..SWC = NA,
-        ..confint = NA,
-        ..desc = NA)
-)
-
-dataMBIcorrResults <- if (requireNamespace('jmvcore')) R6::R6Class(
-    inherit = jmvcore::Group,
-    active = list(
-        text = function() private$.items[["text"]],
-        mbi = function() private$.items[["mbi"]],
-        desc = function() private$.items[["desc"]]),
-    private = list(),
-    public=list(
-        initialize=function(options) {
-            super$initialize(
-                options=options,
-                name="",
-                title="MBI for Correlation")
-            self$add(jmvcore::Preformatted$new(
-                options=options,
-                name="text",
-                title="MBI for Correlation"))
-            self$add(jmvcore::Table$new(
-                options=options,
-                name="mbi",
-                title="Magnitude Based Inference: Correlation",
-                rows="(pairs)",
-                clearWith=list(
-                    "confint",
-                    "SWC"),
-                columns=list(
-                    list(
-                        `name`="i1", 
-                        `title`="", 
-                        `type`="text", 
-                        `content`="($key)"),
-                    list(
-                        `name`="i2", 
-                        `title`="", 
-                        `type`="text", 
-                        `content`="($key)"),
-                    list(
-                        `name`="b[0]", 
[…]
