### 2018-06-13 ffff5b0
Delete datambinpind.b.R
