+++ mbir/R/datambinppair.b.R
@@ -1,263 +1,0 @@
-
-# This file is a generated template, your changes will not be overwritten
-
-dataMBInppairClass <- if (requireNamespace('jmvcore')) R6::R6Class(
-    "dataMBInppairClass",
-    inherit = dataMBInppairBase,
-    private = list(
-      .init = function(){
-        
-        table <- self$results$ttest
-        EffSize <- self$results$effect
-        MBI <- self$results$mbi
-        desc <- self$results$desc
-        
-        
-        
-        for (pair in self$options$pairs) {
-          table$setRow(rowKey=pair,  list(i1=pair[[1]], i2=pair[[2]]))
-          EffSize$setRow(rowKey=pair, list(i1=pair[[1]], i2=pair[[2]]))
-          MBI$setRow(rowKey=pair, list(i1=pair[[1]], i2=pair[[2]]))
-          desc$setRow(rowKey=pair, list(`name[1]`=pair[[1]], `name[2]`=pair[[2]]))
-        }
-        
-        EffSize$getColumn('rLL')$setSuperTitle(jmvcore::format('{}% Confidence Interval', self$options$confint))
-        EffSize$getColumn('rUL')$setSuperTitle(jmvcore::format('{}% Confidence Interval', self$options$confint))          
-        
-        
-      },
-        .run = function() {
-          
-          table <- self$results$ttest
-          EffSize <- self$results$effect
-          MBI <- self$results$mbi
-          desc <- self$results$desc
-          
-          conf.int <- (self$options$confint)/100
-          SWC <- self$options$SWC
-          
-          for (pair in self$options$pairs) {
-            
-            if (is.null(pair[[1]]))
-              next()
-            if (is.null(pair[[2]]))
-              next()    
-            
-            #Gather descriptives
-            i1 <- jmvcore::toNumeric(self$data[[pair[[1]] ]])
-            i2 <- jmvcore::toNumeric(self$data[[pair[[2]] ]])
-            data <- data.frame(i1=i1, i2=i2)
-            data <- na.omit(data)
-            n <- nrow(data)
-            i1 <- data$i1
-            i2 <- data$i2
-            m1 <- base::mean(i1)
-            m2 <- base::mean(i2)
-            med1 <- stats::median(i1)
-            med2 <- stats::median(i2)
-            sd1  <- stats::sd(i1)
-            sd2  <- stats::sd(i2)
-            se1  <- sd1/sqrt(n)
-            se2  <- sd2/sqrt(n)
-           
-            rank <- stats::wilcox.test(x=i1, y=i2, 
-                                       paired = TRUE,
-                                       conf.int = T,
-                                       conf.level = conf.int,
-                                       na.action = na.omit,
-                                       correct = F,
-                                       exact = F)  
-            suppressWarnings(warning(rank))
-            diff <- rank$estimate
-            
-            r <- ifelse(diff < 0,
-                        stats::qnorm(rank$p.value / 2) / sqrt(n),
-                        abs(stats::qnorm(rank$p.value / 2) / sqrt(n)))
-            
-            r.LL <-( (exp(2 * ((SWC * log((1 + r) / (1 - r)
-            )) + (
-              stats::qnorm(((
-                100 -
-                  (100 * conf.int)
-              ) / 100 / 2)) / sqrt(n - 3)
-            ))) - 1) / (exp(2 *
-                              ((SWC * log((1 + r) /
-                                              (1 - r)
-                              )) + (
-                                stats::qnorm(((
-                                  100 -
-                                    (100 * conf.int)
-                                ) / 100 / 2)) / sqrt(n - 3)
-                              ))) + 1) )
-            
-            rLL <- as.numeric(r.LL)
-            
-            r.UL <- ((exp(2 * ((SWC * log((1 + r) / (1 - r)
-            )) - (
[…]
