### 2018-06-13 834b247
Delete datambipaired.b.R
