+++ mbir/R/datambipaired.b.R
@@ -1,294 +1,0 @@
-
-# This file is a generated template, your changes will not be overwritten
-
-dataMBIpairedClass <- if (requireNamespace('jmvcore')) R6::R6Class(
-    "dataMBIpairedClass",
-    inherit = dataMBIpairedBase,
-    private = list(
-        .init = function(){
-          
-          table <- self$results$ttest
-          EffSize <- self$results$effect
-          MBI <- self$results$mbi
-          desc <- self$results$desc
-          
-
-          
-          for (pair in self$options$pairs) {
-            table$setRow(rowKey=pair,  list(i1=pair[[1]], i2=pair[[2]]))
-            EffSize$setRow(rowKey=pair, list(i1=pair[[1]], i2=pair[[2]]))
-            MBI$setRow(rowKey=pair, list(i1=pair[[1]], i2=pair[[2]]))
-            desc$setRow(rowKey=pair, list(`name[1]`=pair[[1]], `name[2]`=pair[[2]]))
-          }
-          
-          EffSize$getColumn('diffLL')$setSuperTitle(jmvcore::format('{}% Confidence Interval', self$options$confint))
-          EffSize$getColumn('diffUL')$setSuperTitle(jmvcore::format('{}% Confidence Interval', self$options$confint)) 
-          EffSize$getColumn('dLL')$setSuperTitle(jmvcore::format('{}% Confidence Interval', self$options$confint))
-          EffSize$getColumn('dUL')$setSuperTitle(jmvcore::format('{}% Confidence Interval', self$options$confint))          
-          
-          
-        },
-        .run = function() {
-
-          
-          table <- self$results$ttest
-          EffSize <- self$results$effect
-          MBI <- self$results$mbi
-          desc <- self$results$desc
-          
-          conf.int <- (self$options$confint)/100
-          SWC <- self$options$SWC
-          
-          for (pair in self$options$pairs) {
-            
-            if (is.null(pair[[1]]))
-              next()
-            if (is.null(pair[[2]]))
-              next()
-            
-            #Gather descriptives
-            i1 <- jmvcore::toNumeric(self$data[[pair[[1]] ]])
-            i2 <- jmvcore::toNumeric(self$data[[pair[[2]] ]])
-            data <- data.frame(i1=i1, i2=i2)
-            data <- na.omit(data)
-            n <- nrow(data)
-            i1 <- data$i1
-            i2 <- data$i2
-            m1 <- base::mean(i1)
-            m2 <- base::mean(i2)
-            med1 <- stats::median(i1)
-            med2 <- stats::median(i2)
-            sd1  <- stats::sd(i1)
-            sd2  <- stats::sd(i2)
-            se1  <- sd1/sqrt(n)
-            se2  <- sd2/sqrt(n)
-            
-            result <- t.test(i1, i2, paired=TRUE)
-            corr <- cor.test(x=i1,y=i2)
-            corr <- unname(corr$estimate)
-            
-            #Mean Difference
-            mdiff <- m1-m2
-            diffSE <- sqrt(((sd1^2/n)+(sd2^2/n))-(2*corr*(sd1/sqrt(n))*(sd2/sqrt(n))))
-            diffLL <- mdiff+diffSE*qt((1-conf.int)/2,n)
-            diffUL <- mdiff-diffSE*qt((1-conf.int)/2,n)
-            
-            #Effect Size
-            d <- (result$statistic/sqrt(n))
-            
-            LL <- d - (stats::qt(((
-              100 - (100 * conf.int)
-            ) / 100) / 2,
-            result$parameter)) * abs(d) / stats::qt(result$p.value /
-                                                      2,
-                                                    result$parameter)
-            
[…]
