+++ mbir/R/datambipaired.h.R
@@ -1,387 +1,0 @@
-
-# This file is automatically generated, you probably don't want to edit this
-
-dataMBIpairedOptions <- if (requireNamespace('jmvcore')) R6::R6Class(
-    "dataMBIpairedOptions",
-    inherit = jmvcore::Options,
-    public = list(
-        initialize = function(
-            pairs = NULL,
-            correl = TRUE,
-            SWC = 0.5,
-            confint = 90,
-            desc = FALSE, ...) {
-
-            super$initialize(
-                package='mbir',
-                name='dataMBIpaired',
-                requiresData=TRUE,
-                ...)
-
-            private$..pairs <- jmvcore::OptionPairs$new(
-                "pairs",
-                pairs,
-                suggested=list(
-                    "continuous"),
-                permitted=list(
-                    "continuous",
-                    "nominal",
-                    "ordinal"))
-            private$..correl <- jmvcore::OptionBool$new(
-                "correl",
-                correl,
-                default=TRUE)
-            private$..SWC <- jmvcore::OptionNumber$new(
-                "SWC",
-                SWC,
-                min=0,
-                default=0.5)
-            private$..confint <- jmvcore::OptionNumber$new(
-                "confint",
-                confint,
-                min=50,
-                max=99.9,
-                default=90)
-            private$..desc <- jmvcore::OptionBool$new(
-                "desc",
-                desc,
-                default=FALSE)
-
-            self$.addOption(private$..pairs)
-            self$.addOption(private$..correl)
-            self$.addOption(private$..SWC)
-            self$.addOption(private$..confint)
-            self$.addOption(private$..desc)
-        }),
-    active = list(
-        pairs = function() private$..pairs$value,
-        correl = function() private$..correl$value,
-        SWC = function() private$..SWC$value,
-        confint = function() private$..confint$value,
-        desc = function() private$..desc$value),
-    private = list(
-        ..pairs = NA,
-        ..correl = NA,
-        ..SWC = NA,
-        ..confint = NA,
-        ..desc = NA)
-)
-
-dataMBIpairedResults <- if (requireNamespace('jmvcore')) R6::R6Class(
-    inherit = jmvcore::Group,
-    active = list(
-        text = function() private$.items[["text"]],
-        ttest = function() private$.items[["ttest"]],
-        effect = function() private$.items[["effect"]],
-        mbi = function() private$.items[["mbi"]],
-        desc = function() private$.items[["desc"]]),
-    private = list(),
-    public=list(
-        initialize=function(options) {
-            super$initialize(
-                options=options,
-                name="",
-                title="MBI for Paired Samples")
-            self$add(jmvcore::Preformatted$new(
-                options=options,
-                name="text",
-                title="MBI for Paired Samples"))
-            self$add(jmvcore::Table$new(
-                options=options,
-                name="ttest",
-                title="Paired Samples T-Test",
-                rows="(pairs)",
-                clearWith=list(
-                    "confint",
-                    "SWC"),
-                columns=list(
-                    list(
-                        `name`="i1", 
-                        `title`="", 
-                        `type`="text", 
-                        `content`="($key)"),
-                    list(
-                        `name`="i2", 
-                        `title`="", 
-                        `type`="text", 
-                        `content`="($key)"),
-                    list(
-                        `name`="t", 
-                        `type`="number"),
-                    list(
-                        `name`="df", 
-                        `type`="integer"),
-                    list(
-                        `name`="p", 
-                        `type`="number", 
-                        `format`="zto,pvalue"))))
-            self$add(jmvcore::Table$new(
-                options=options,
-                name="effect",
-                title="Effect Size",
-                rows="(pairs)",
-                clearWith=list(
[…]
