### 2018-06-13 e43abdd
Delete datambitwo.b.R
