### 2018-06-13 c29e4df
Delete datambitwo.h.R
