+++ README.md
@@ -1,5 +1,6 @@
 [![version](http://www.r-pkg.org/badges/version/mbir)](https://cran.r-project.org/package=mbir)
 [![CRAN](https://cranlogs.r-pkg.org/badges/mbir)](https://cran.r-project.org/web/packages/mbir/index.html)
+[![Build Status](https://travis-ci.org/kdpeterson51/mbir.svg?branch=master)](https://travis-ci.org/kdpeterson51/mbir)
 
 # mbir: Magnitude-Based Inferences
 Kyle D Peterson and Aaron R Caldwell
