### 2018-06-13 kdpeterson51/mbir#2 @kdpeterson51
@kellieotto, thank you for your patience. mbir has successfully [![Build Status](https://travis-ci.org/kdpeterson51/mbir.svg?branch=master)](https://travis-ci.org/kdpeterson51/mbir) with Travis CI.

Please inform me if there is anything you may need from me.

Again, thank you.
