### 2018-06-13 #746 @kdpeterson51
@leeper Both issues raised by @kellieotto have now been addressed.

Please inform me if there is anything else you may need form me. Thank you very much.
