### 2018-06-13 kdpeterson51/mbir#2 @kellieotto
Great, Travis CI looks like it's working!

Your package still needs comprehensive unit tests for its functions. I'll point you to some resources on how to do this. Please let me know if you need more information.

http://r-pkgs.had.co.nz/tests.html
https://journal.r-project.org/archive/2011-1/RJournal_2011-1_Wickham.pdf
https://testthat.r-lib.org/index.html
