### 2018-06-13 #746 @kellieotto
The version issue has been addressed. The package still needs unit tests.
