### 2018-06-13 #746 @leeper
Yes, please add unit tests as suggested here: https://github.com/kdpeterson51/mbir/issues/2
