+++ mbir/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(mbir)
+
+test_check("mbir")
