### 2018-06-22 kdpeterson51/mbir#2 @kellieotto
Good start. I hate to nitpick but these unit tests aren't comprehensive. They check that your functions are internally coherent with each other, but only for one example each. I'd like to see tests for more (ideally all) of your functions, more than one test for each, and testing that the outputs of your functions match some external reference (i.e. not just that they match each other).
