### 2018-06-22 kdpeterson51/mbir#2 @leeper
I agree.
