### 2018-06-26 be68af3
Updated ttesthat functions; addition of external validation (still in progress)
