+++ mbir/R/smd.R
@@ -112,5 +112,6 @@ smd <- function (es, p, df, conf.int=0.9, swc=0.5, plot=FALSE)
 
   rval <- list(es=es, es.LL=LL, es.UL=UL,
                p.value=p, conf.int=conf.int, swc=swc,
-               mbiPositive=positive, mbiTrivial=trivial, mbiNegative=negative, inference=inference)
+               mbiPositive=positive, mbiTrivial=trivial,
+               mbiNegative=negative, inference=inference)
 }
