### 2018-06-26 a64f120
Modified context statement
