+++ mbir/NAMESPACE
@@ -1,5 +1,5 @@
+# Generated by roxygen2: do not edit by hand
 
-importFrom("stats", "na.omit", "sd")
 export(aipe_smd)
 export(boot_test)
 export(corr)
