+++ mbir/DESCRIPTION
@@ -1,7 +1,7 @@
 Package: mbir
 Type: Package
 Title: Magnitude-Based Inferences
-Version: 1.3.1
+Version: 1.3.2
 Date: 2018-04-28
 Authors@R: c(person("Kyle", "Peterson", 
                      email = "petersonkdon@gmail.com",
@@ -9,12 +9,7 @@ Authors@R: c(person("Kyle", "Peterson",
              person("Aaron", "Caldwell", role = "aut"))
 Maintainer: Kyle Peterson <petersonkdon@gmail.com>
 Description: Allows practitioners and researchers a wholesale approach for deriving magnitude-based inferences from raw data. A major goal of 'mbir' is to programmatically detect appropriate statistical tests to run in lieu of relying on practitioners to determine correct stepwise procedures independently.
-Imports: 
-  graphics, 
-  stats, 
-  utils, 
-  jmvcore (>= 0.8),
-  R6
+Imports: graphics, stats, utils
 URL: http://mbir-project.us/
 License: GPL-2
 Copyright: Segments of the package are based upon Will G. Hopkins' work. See vignette and COPYRIGHT file for details.
@@ -22,6 +17,5 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
 Suggests: knitr,
-    rmarkdown,
-    testthat
+    rmarkdown
 VignetteBuilder: knitr
