### 2018-07-11 #746 @leeper
@kdpeterson51 I see you've started adding tests. Let me know when you feel the package is ready for me to take another, more thorough look. Thanks!
