### 2018-07-11 cc4f33e
Update corr_test.R
