### 2018-07-11 789e9af
Update smd_test.R
