### 2018-07-11 248d26f
Update swc_ind.R
