### 2018-07-11 34d7940
Update corr_test.R
