### 2018-07-11 538b0ca
Update smd_test.R
