### 2018-07-11 dfe6146
Update DESCRIPTION
