+++ mbir/DESCRIPTION
@@ -17,5 +17,6 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
 Suggests: knitr,
+    testthat,
     rmarkdown
 VignetteBuilder: knitr
