### 2018-07-11 545d556
Add requirement to description for effsize.
