### 2018-07-11 3e650ad
edit testthat functions
