+++ mbir/tests/testthat.R
@@ -1,4 +1,5 @@
 library(testthat)
 library(mbir)
+library(effsize)
 
 test_check("mbir")
