### 2018-07-11 738c6f7
modify testthat again.
