### 2018-07-11 298f29a
Well, now things should be fixed. Errors are my fault in the testhat functions.
Cheers,

Aaron
