### 2018-07-11 8c34ec7
smd corr fixes
