### 2018-07-11 95b1782
smd corr vs corr test fixed
