+++ mbir/R/smd.R
@@ -13,8 +13,8 @@
 #'@examples smd(.75, 0.06, 20, 0.95)
 #'@export
 
-smd <- function (es, p, df, conf.int=0.9, swc=0.5, plot=FALSE)
-{
+smd <- function (es, p, df, conf.int=0.9, swc=0.5, plot=FALSE){
+
   if (is.character(es) == TRUE || is.factor(es) == TRUE ||
       is.character(p) == TRUE || is.factor(p) == TRUE || is.character(df) ==
       TRUE || is.factor(df) == TRUE) {
@@ -37,13 +37,13 @@ smd <- function (es, p, df, conf.int=0.9, swc=0.5, plot=FALSE)
     stop(error)
   }
 
-  negative <- round(100 * (ifelse((es - -0.5) > 0, stats::pt((es -
-                                                                -0.5)/abs(es) * abs(stats::qt(p/2, df)), df, lower.tail = F),
-                                  (1 - stats::pt((-0.5 - es)/abs(es) * abs(stats::qt(p/2,
+  negative <- round(100 * (ifelse((es - -swc) > 0, stats::pt((es -
+                                                                -swc)/abs(es) * abs(stats::qt(p/2, df)), df, lower.tail = F),
+                                  (1 - stats::pt((-swc - es)/abs(es) * abs(stats::qt(p/2,
                                                                                      df)), df, lower.tail = F)))), digits = 1)
-  positive <- round(100 * (ifelse((es - 0.5) > 0, (1 - stats::pt((es -
-                                                                    0.5)/abs(es) * abs(stats::qt(p/2, df)), df, lower.tail = F)),
-                                  stats::pt((0.5 - es)/abs(es) * abs(stats::qt(p/2, df)),
+  positive <- round(100 * (ifelse((es - swc) > 0, (1 - stats::pt((es -
+                                                                    swc)/abs(es) * abs(stats::qt(p/2, df)), df, lower.tail = F)),
+                                  stats::pt((swc - es)/abs(es) * abs(stats::qt(p/2, df)),
                                             df, lower.tail = F))), digits = 1)
   trivial <- round((100 - positive - negative), digits = 1)
   LL <- es - (stats::qt(((100 - (100 * conf.int))/100)/2, df)) *
@@ -59,17 +59,17 @@ smd <- function (es, p, df, conf.int=0.9, swc=0.5, plot=FALSE)
   table <- matrix(c("Negative", "Trivial", "Positive", negative,
                     trivial, positive), nrow = 2, byrow = T)
   rownames(table) <- c(" ", "MBI (%)")
-  lower <- ifelse(negative < 0.5, "Most Unlikely", ifelse(negative <
+  lower <- ifelse(negative < swc, "Most Unlikely", ifelse(negative <
                                                             5, "Very Unlikely", ifelse(negative < 25, "Unlikely",
                                                                                        ifelse(negative < 75, "Possibly", ifelse(negative < 95,
                                                                                                                                 "Likely", ifelse(negative < 99, "Most Likely", ifelse(negative >=
                                                                                                                                                                                         99, "Almost Certainly")))))))
-  trivial2 <- ifelse(trivial < 0.5, "Most Unlikely", ifelse(trivial <
+  trivial2 <- ifelse(trivial < swc, "Most Unlikely", ifelse(trivial <
                                                               5, "Very Unlikely", ifelse(trivial < 25, "Unlikely",
                                                                                          ifelse(trivial < 75, "Possibly", ifelse(trivial < 95,
                                                                                                                                  "Likely", ifelse(negative < 99, "Most Likely", ifelse(negative >=
[…]

[…]
