### 2018-07-13 85a4b23
Update corr_test.R
