### 2018-07-13 f1b43d0
Update smd_test.R
