+++ mbir/tests/testthat/test-ES-equivalent.R
@@ -1,6 +1,5 @@
 context("Do the corr_test and smd_test function calculate proper effect sizes")
 library(mbir)
-library(effsize)
 
 data <- read.csv("https://raw.githubusercontent.com/jasp-stats/jasp-desktop/development/Resources/Data%20Sets/Big%205%20(Dolan%2C%20Oort%2C%20Stoel%20%26%20Wicherts%2C%202009).csv", sep="")
 #data <- read.csv("C:/Users/Daniel/Downloads/Big 5 (Dolan, Oort, Stoel & Wicherts, 2009).csv", sep="")
