### 2018-07-13 6a22ec4
Update test-sample-size.R
