### 2018-07-13 #746 @kdpeterson51
@leeper @kellieotto we have added more comprehensive unit tests and would like your feedback on the matter. Thank you for all that you do.
