### 2018-07-13 7c7f86a
Update test-sample-size.R
