### 2018-08-01 kdpeterson51/mbir#2 @kellieotto
I see that you've added some new tests. These look great. Again, I'd like to see tests for more (ideally all) of your functions and testing that the outputs of your functions match some external reference.
