### 2018-08-01 #746 @kellieotto
I'm very sorry for such a long delay. This got buried in my inbox. The tests are a good start, but I'd still like to see more comprehensive ones. [See my comment.](https://github.com/kdpeterson51/mbir/issues/2)

Might be useful to have a look at [this resource or other tutorials on unit testing.](https://katyhuff.github.io/python-testing/) This tutorial is Python specific, but there are some good conceptual things in there that apply to R packages too.
