### 2018-08-04 kdpeterson51/mbir#2 @leeper
Just to keep track, I see tests of the following:

- [x] aipe_smd
- [x] boot_test
- [x] corr
- [ ] corr_diff
- [x] corr_test
- [ ] es_convert
- [ ] odds
- [ ] prop
- [x] smd
- [x] smd_test
- [x] ss_corr
- [x] ss_odds
- [x] ss_smd
- [ ] swc_ind

It would be good to see tests of the remaining functions.
