### 2018-08-04 1c88c43
Update prop.R
