### 2018-08-26 kdpeterson51/mbir#2 @leeper
@kdpeterson51 Just a ping. It would be great to see this final issue of testing resolved, as we're very close to being able to accept this.
