### 2018-09-16 3fb0afd
adding test that for diff_corr function.
