+++ mbir/DESCRIPTION
@@ -2,20 +2,20 @@ Package: mbir
 Type: Package
 Title: Magnitude-Based Inferences
 Version: 1.3.4
-Date: 2018-07-12
+Date: 2018-09-16
 Authors@R: c(person("Kyle", "Peterson", 
                      email = "petersonkdon@gmail.com",
                      role = c("aut", "cre")),
              person("Aaron", "Caldwell", role = "aut"))
 Maintainer: Kyle Peterson <petersonkdon@gmail.com>
 Description: Allows practitioners and researchers a wholesale approach for deriving magnitude-based inferences from raw data. A major goal of 'mbir' is to programmatically detect appropriate statistical tests to run in lieu of relying on practitioners to determine correct stepwise procedures independently.
-Imports: graphics, stats, utils, effsize
+Imports: graphics, stats, utils, effsize, psych
 URL: http://mbir-project.us/
 License: GPL-2
 Copyright: Segments of the package are based upon Will G. Hopkins' work. See vignette and COPYRIGHT file for details.
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 6.0.1
+RoxygenNote: 6.1.0
 Suggests: knitr,
     testthat,
     rmarkdown
