### 2018-09-17 3fad9f8
Update test-data_summary_equivalent.R
