### 2018-09-17 b1ed146
Update test-data_summary_equivalent.R
