+++ mbir/inst/doc/mbir-vignette.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Introduction to `mbir`"
-author: "Kyle D. Peterson"
+author: "Kyle D. Peterson, Aaron Caldwell"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
@@ -89,7 +89,7 @@ In a nutshell, when a Gaussian distribution is detected, the terminal effect siz
 
 - Unfortunately, *d* overestimates size of effects with small sample sizes. Consequently, to reduce bias and avoid overestimating *d*, an approximate adjustment factor is applied to *d* when degrees of freedom are below 30.^[Hedges LV. (1981). Distribution theory for Glass’s estimator of effect size and related estimators. *Journal of Educational Statistics*, 6, 107–128] Thereafter, *d*~adj~ represents the adjusted effect size value. 
 
-Upon log-transformation, exponentiating the confidence intervals of the log-transformed difference consequently bestows a ratio of geometric means.^[Altman DG. (1991). *Practical statistics for medical research*. London, UK: Chapman and Hall] Subsequently taking the difference between these two values captures the point estimate - that is, on the log scale: log(a)-log(b) = log(a/b), which can be interpreted in a similar manner as an odds ratio. On the nonparametric side, dividing the *z* value by the square root of *n* stealthily arrives at *r*. 
+Upon log-transformation, exponentiating the confidence intervals of the log-transformed difference consequently bestows a difference in geometric means.^[Altman DG. (1991). *Practical statistics for medical research*. London, UK: Chapman and Hall] Subsequently taking the difference between these two values captures the point estimate - that is, on the log scale: log(a)-log(b) = log(a/b), which can be interpreted in a similar manner as an odds ratio. On the nonparametric side, dividing the *z* value by the square root of *n* stealthily arrives at *r*. 
 
 - Effect size confidence intervals are provided for each analysis to disclose the precision of observed effect size, or perhaps the maximum likely estimation error.   
 
@@ -97,8 +97,7 @@ Upon log-transformation, exponentiating the confidence intervals of the log-tran
 
 The final calculations aggregate the aforementioned statistical parameters to ultimately formulate magnitude-based inferences of effects. Calculations were reverse-engineered from Hopkins's cited spreadsheet^[Hopkins WG. (2007). A spreadsheet for deriving a confidence interval, mechanistic inference and clinical inference from a *p* value. *Sportscience* 11, 16-20. sportsci.org/2007/wghinf.htm.] and transcribed into the **R** language. Users are welcome to use `mbir` Input Functions (Section 2) to directly compare their numerical results to the respective spreadsheet and thereby confirm `mbir` calculations are correct. 
 
-To quickly abbreviate, magnitude-based inferences offer a form of inference in which the parameter is partitioned into three regions corresponding to various qualitative
-aspects of the parameter (substantially negative, trivial, substantially positive). A probability is assigned to each of these regions, calculated from the assumed distribution of the parameter estimate. To be more transparent, the obtained *t* statistic is adjusted (updated) by the calculated effect size before justifying the cumulative probability distribution. (This marvelous statistical modulation has therefore been coined "pseudo-Bayesian"). Each `mbir` function presents the observed magnitude of effect, as well as the precision of estimation. *P*-values are shown for the sake of formality, but users are encouraged to highlight Hopkins' content-rich inference: making a qualitative statement with regard to practical significance. The last line of `smd_test()` output composes the appropriate statement from given parameters, which considers both the level of confidence (*possibly*, *likely*, etc), level of magnitude (trivial, small, moderate, etc), as well as the direction (negative, positive) of effect.
[…]
