+++ mbir/R/swc_ind.R
@@ -78,7 +78,7 @@ swc_ind<-function(x, swc, type=c("previous","trend"), ts, te, main, xlab, ylab){
     ylab<-deparse(substitute(x))
   }
 
-  times<<-c(1:length(x))
+  times<-c(1:length(x))
   mylen<-times[-utils::tail(times,n=1)]
   times2<-c(2:length(x))
 
@@ -90,16 +90,34 @@ swc_ind<-function(x, swc, type=c("previous","trend"), ts, te, main, xlab, ylab){
   point2<-offset$diffn[-1]
   prev<-offset$x[1:length(offset$x)-1]
 
-  mod<-stats::lm(x~times)
-  sum.lm<-stats::summary.lm(mod)
-  pred<-stats::predict.lm(mod,newdata=1,se.fit = T,interval = "confidence")
+  X <- cbind(1, times)
+  colnames(X)[1] <- "Intercept"
+  y <- x
+  A <- base::solve(t(X) %*% X)
+  beta <- (A %*% t(X) %*% y)
+  slope <- beta[[2]]
+  resid <- (y - X %*% beta)
+  fit <- (y - resid)
+  residual_var <- base::sum(resid^2) / (length(x) - 1 - 1)
+  residual.scale <- sqrt(residual_var)
+  rss<-base::sum(resid^2)
+  mss <- base::sum((fit - mean(fit))^2)
+  r.squared <- mss/(mss + rss)
+  adj.r.squared <- 1 - (1 - r.squared) * ((length(x) - 1)/(length(x) - 1 - 1))
+  df.res <- (length(x) - 1 - 1)
+  beta_covar <- (residual_var * A)
+  beta_SE <- base::sqrt(base::diag(beta_covar))
+  T.stat <- slope/beta_SE
+  p.value <- 2 * stats::pt(-base::abs(T.stat), length(x)-2)
+  tha.p.value <- p.value[[2]]
+  f.val <- (mss/(1))/residual_var
 
-  trivband.plus<-mod$fitted.values+swc*mod$fitted.values/100+pred$residual.scale*sqrt((1/length(times))+(1/(length(times)-1))*(times-mean(times))^2/stats::sd(times)^2)
-  trivband.minus<-mod$fitted.values-swc*mod$fitted.values/100-pred$residual.scale*sqrt((1/length(times))+(1/(length(times)-1))*(times-mean(times))^2/stats::sd(times)^2)
+  trivband.plus<-fit+swc*fit/100+residual.scale*sqrt((1/length(times))+(1/(length(times)-1))*(times-mean(times))^2/stats::sd(times)^2)
+  trivband.minus<-fit-swc*fit/100-residual.scale*sqrt((1/length(times))+(1/(length(times)-1))*(times-mean(times))^2/stats::sd(times)^2)
 
 
   if(missing(te)){
-    te<-pred$residual.scale
+    te<-residual.scale
   } else {
     te<-te
     }
@@ -133,24 +151,23 @@ swc_ind<-function(x, swc, type=c("previous","trend"), ts, te, main, xlab, ylab){
   graphics::lines(times,trivband.plus,lty=2,col="#666666")
   graphics::lines(times,trivband.minus,lty=2,col="#666666")
   graphics::arrows(times, x-te, times, x+te, length=0.05, angle=90, code=3)
-  rm(times,envir = globalenv())
   }
 
 
   if(type == "trend"){
 
-  pred.err<-sqrt(te^2+pred$residual.scale^2*((1/length(times)+1/(length(times)-1))*(times-mean(times))^2/sd(times)^2))
[…]
