+++ mbir/DESCRIPTION
@@ -2,7 +2,7 @@ Package: mbir
 Type: Package
 Title: Magnitude-Based Inferences
 Version: 1.3.5
-Date: 2019-01-17
+Date: 2019-01-21
 Authors@R: c(person("Kyle", "Peterson", 
                      email = "petersonkdon@gmail.com",
                      role = c("aut", "cre")),
@@ -20,6 +20,6 @@ RoxygenNote: 6.1.0
 Suggests: knitr, testthat, rmarkdown
 VignetteBuilder: knitr
 NeedsCompilation: no
-Packaged: 2019-01-18 04:05:32 UTC; ath_sportsci
+Packaged: 2019-01-22 04:08:33 UTC; ath_sportsci
 Author: Kyle Peterson [aut, cre],
   Aaron Caldwell [aut]
