### 2019-01-22 kdpeterson51/mbir#2 @kellieotto
Looks like the functions with unchecked boxes can be checked off now.
