### 2019-01-22 #746 @kellieotto
@arfon @kdpeterson51 Looks like the functions with missing unit tests have them now. I've closed the issue on it. Great! I recommend to accept.
